Foi desenvolvido uma ambiente de produção e desenvolvimento para melhor aproveitamento da equipe

+++ src/main/java/com/api/repository/ItemPedidoRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.api.domain.ItemPedido;
 
 @Repository
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer>{
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long>{
 
 }
