+++ src/main/java/com/api/repository/PedidoRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.api.domain.Pedido;
 
 @Repository
-public interface PedidoRepository extends JpaRepository<Pedido, Integer>{
+public interface PedidoRepository extends JpaRepository<Pedido, Long>{
 
 }
